move navigation out of core
